cursora/requirements.txt
apparmor==3.0.1
appdirs==1.4.4
asgiref==3.3.1
asn1crypto==1.4.0
Beaker==1.11.0
CacheControl==0.12.6
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
distlib==0.3.1
distro==1.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
LibAppArmor==3.0.1
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
msgpack==1.0.2
oauthlib==3.1.0
ordered-set==4.0.2
packaging==20.9
pep517==0.9.1
Pillow==8.1.0
ply==3.11
progress==1.5
psycopg2==2.8.6
pycparser==2.20
PyGObject==3.38.0
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
resolvelib==0.5.4
retrying==1.3.3
six==1.15.0
slugify==0.0.1
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
team==1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
webencodings==0.5.1
